+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,12 +1,10 @@
 package lotto.domain.lotto;
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
-import lotto.domain.primitive.LottoNumber;
 
 public class Lotto {
