+++ src/main/java/lotto/domain/lotto/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.domain.primitive;
+package lotto.domain.lotto;
 
 import java.util.Objects;
