+++ src/main/java/lotto/domain/manager/LottoManager.java
@@ -1,9 +1,8 @@
-package lotto.domain;
+package lotto.domain.manager;
 
 import lotto.domain.lotto.Lotto;
 import lotto.domain.lotto.LottoRepository;
 import lotto.domain.lottomachine.LottoMachine;
-import lotto.domain.primitive.Ticket;
 import lotto.domain.rating.Rating;
 import lotto.domain.rating.RatingInfo;
 import lotto.domain.statistics.WinningLotto;
