+++ src/test/java/lotto/domain/lotto/LottoNumberTest.java
@@ -1,7 +1,8 @@
-package lotto.domain.primitive;
+package lotto.domain.lotto;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.domain.lotto.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
